+++ x86/subops/x86_0F_BA.c
@@ -2,7 +2,7 @@
 
 void ASM_0F_BABT(RM *rm, s32 disp, u8 val) {
     if (rm->isPtr) {
-        s64 fdisp = ASM_getDisp(rm, disp);
+        u32 fdisp = ASM_getDisp(rm, disp);
         
         switch (rm->otype) {
             case R_Bit16: { STACK16(temp, fdisp); f.f.cf = ((*temp & (1 << val)) == (1 << val)); break; }
